feat: add global usepagination
